[TASK] Gateways: Use common superclass (RT-3215)
And set noRipple flag on USD gateway.
